Move asset conversion utilities into a shared class.
This allows compilers for different asset types to share implementations
of converting asset information into compiled representations.
